backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
web3==6.11.3
python-dotenv==1.0.0
Werkzeug==2.3.7
py-solc-x==1.1.1
cryptography==41.0.7

